rpm: make _generate_package_note_file always recreate the file
The idea was that we can avoid unnecessary work if the macro is called more than
once. But in hindsight this might be risky: let's instead minimize the number of
places where the macros is called, but always overwrite the file so that we
don't end up with a stale version from a previous build.

Cherry-picked from https://src.fedoraproject.org/rpms/package-notes/c/1acac701a662914ecf1542df853bed6b7b58d228