dlopen-notes: add mode to generate rpm Provides/Requires
This can be used to insert the appropriate deps into the spec file:
$ ./dlopen-notes.py .../libsystemd-shared-256.so --rpm-requires zstd,qrencode --rpm-recommends tpm
Requires: libqrencode.so.4()(64bit)
Requires: libzstd.so.1()(64bit)
Recommends: libtss2-esys.so.0()(64bit)
Recommends: libtss2-rc.so.0()(64bit)
Recommends: libtss2-mu.so.0()(64bit)

Most likely, it needs to be used manually. Rpm doesn't really have a
notion of dynamically generated deps at the package level :(

(There is the fileattr mechanism,
https://rpm-software-management.github.io/rpm/manual/dependency_generators.html,
but that is per-file. It has the shortcoming that it assumes that the
generated deps depend only on the file itself, and not the subpackage that the
file lands in, and there is no notion of package configuration. We _could_ use
that, but it seems a poor fit.

Also, before rpm 4.20, fileattr generators could not be provided by the
package being built, they needed to be installed in the filesystem. I don't
think we're at the point where it makes sense to make this "global".

4.15 added %generate_buildrequires. Obviously that only works for
BuildRequires. Also, the mechanism with generating src.rpms in a loop is
rather convoluted and heavyweight.)

I'll open an RFE in rpm. For now, the generator can be used manually in
systemd. That is still nicer than the previous method of looking at the
sources.