dh_dlopenlibdeps: Debian package names can contain dots
According to Debian Policy §5.6.1 and 5.6.7, all other characters that
are allowed in a Debian package name were already covered by the regex
used here, but the dot was missing.

For example this fixes the ability to generate dependencies from
ELF notes that point to libsndio.so.7 (libsndio7.0 package) or
libpipewire-0.3.so.0 (libpipewire-0.3-0 package).

Closes: #1117245
Signed-off-by: Simon McVittie <smcv@debian.org>